requirements.txt
aiogram==2.25.2
gspread==5.12.4
google-auth==2.22.0
apscheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1